+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Unreleased
 * Voltage dependent loads.
 * Changed parent class of AttributeInt in order to make it work better with numpy.
 * Load voltage dependence constraint.
+* Unittest for raw write.
   
 Version 1.3.3
 -------------